chore: changes from PR
